Move updateProduct, addNewProduct and deleteProduct fetch calls into apiCalls.js and import them into MainSectionContent.js for use
