+++ db.json
@@ -43,7 +43,7 @@
     {
       "id": 6,
       "img": "/productImages/iphone13.png",
-      "name": "iPhone 13 Pro Rose Gold",
+      "name": "iPhone 143 Pro Rose Gold",
       "sku": "00422001",
       "description": "A super bright display in a durable design. Hollywood-worthy video shooting made easy. A lightning-fast chip. And a big boost in battery life you’ll notice every day.",
       "price": 200,
@@ -53,7 +53,7 @@
     {
       "id": 7,
       "img": "/productImages/iphone13.png",
-      "name": "iPhone 13 Pro Rose Gold",
+      "name": "iPhone 133 Pro Rose Gold",
       "sku": "00422001",
       "description": "A super bright display in a durable design. Hollywood-worthy video shooting made easy. A lightning-fast chip. And a big boost in battery life you’ll notice every day.",
       "price": 200,
@@ -61,24 +61,14 @@
       "status": "In stock"
     },
     {
-      "id": 8,
-      "img": "/productImages/iphone13.png",
-      "name": "iPhone 13 Pro Rose Gold",
-      "sku": "00422001",
-      "description": "A super bright display in a durable design. Hollywood-worthy video shooting made easy. A lightning-fast chip. And a big boost in battery life you’ll notice every day.",
-      "price": 200,
-      "quantity": 120,
-      "status": "In stock"
-    },
-    {
-      "id": 9,
-      "img": "/productImages/iphone13.png",
-      "name": "iPhone 89 Pro Rose Gold",
-      "sku": "00422001",
-      "description": "A super bright display in a durable design. Hollywood-worthy video shooting made easy. A lightning-fast chip. And a big boost in battery life you’ll notice every day.",
-      "price": 200,
-      "quantity": 120,
-      "status": "In stock"
+      "img": "https://www.macerkopf.de/wp-content/uploads/2019/04/airpods2_test_8-1.jpg?x18969",
+      "name": "Air Pods 2",
+      "sku": "3893493890843",
+      "description": "Apple Air Pods Gen 2 wireless noise canceling ",
+      "price": "300",
+      "quantity": "22",
+      "status": "On order",
+      "id": 8
     }
   ]
 }
